Integrate Project(Create|Edit) into ProjectEdit
